move Obsoletes from subpackages up to the main one
urpmi can't deal with multiple packages obsoleting the same package, so it 
just doesn't do anything instead of upgrading the package to the newer version.
We need anyway all the subpackages to be installed, so we can put the Obsoletes
on the main package which drags all the subpackages.


SILENT: svn-revision: 467429